+++ NEWS.md
@@ -1,3 +1,11 @@
+# dfidx 0.0-4
+
+* Liming Wang 26 oct 2020, bug for intercept only models fixed in
+  unfold_idx function
+
+* reshape is unsafe with tibbles, coerce to data frame and then back
+  to tibble in this case
+
 # dfidx 0.0-1
 
 * initial version of dfidx posted on github